+++ crates/perry-hir/src/walker.rs
@@ -410,6 +410,34 @@ where
             f(extractable);
             f(usages);
         }
+        Expr::WebCryptoWrapKey {
+            format,
+            key,
+            wrapping_key,
+            wrap_algorithm,
+        } => {
+            f(format);
+            f(key);
+            f(wrapping_key);
+            f(wrap_algorithm);
+        }
+        Expr::WebCryptoUnwrapKey {
+            format,
+            wrapped_key,
+            unwrapping_key,
+            unwrap_algorithm,
+            unwrapped_key_algorithm,
+            extractable,
+            usages,
+        } => {
+            f(format);
+            f(wrapped_key);
+            f(unwrapping_key);
+            f(unwrap_algorithm);
+            f(unwrapped_key_algorithm);
+            f(extractable);
+            f(usages);
+        }
         Expr::CryptoRandomFillSync {
             buffer,
             offset,
@@ -1723,6 +1751,34 @@ where
             usages,
         } => {
             f(algorithm);
+            f(extractable);
+            f(usages);
+        }
+        Expr::WebCryptoWrapKey {
+            format,
+            key,
+            wrapping_key,
+            wrap_algorithm,
+        } => {
+            f(format);
+            f(key);
+            f(wrapping_key);
+            f(wrap_algorithm);
+        }
+        Expr::WebCryptoUnwrapKey {
+            format,
+            wrapped_key,
+            unwrapping_key,
+            unwrap_algorithm,
+            unwrapped_key_algorithm,
+            extractable,
+            usages,
+        } => {
+            f(format);
+            f(wrapped_key);
+            f(unwrapping_key);
+            f(unwrap_algorithm);
+            f(unwrapped_key_algorithm);
             f(extractable);
             f(usages);
         }